minor bug fix in loading weighted graphs

+++ lib/io/graph_io.cpp
@@ -139,7 +139,7 @@ int graph_io::readGraphWeighted(graph_access & G, std::string filename) {
         
         NodeID node_counter   = 0;
         EdgeID edge_counter   = 0;
-        long total_nodeweight = 0;
+        long long total_nodeweight = 0;
 
         G.start_construction(nmbNodes, nmbEdges);
 
@@ -158,7 +158,7 @@ int graph_io::readGraphWeighted(graph_access & G, std::string filename) {
                 if( read_nw ) {
                         ss >> weight;
                         total_nodeweight += weight;
-                        if( total_nodeweight > (long) std::numeric_limits<NodeWeight>::max()) {
+                        if( total_nodeweight > (long long) std::numeric_limits<NodeWeight>::max()) {
                                 std::cerr <<  "The sum of the node weights is too large (it exceeds the node weight type)."  << std::endl;
                                 std::cerr <<  "Currently not supported. Please scale your node weights."  << std::endl;
                                 exit(0);
